switch to musl-clang
unfortunately, there is a known issue with curl/openssl where the resulting (so-called)
static binary throws this error when running on glibc <2.31:

dl-call-libc-early-init.c:37: _dl_call_libc_early_init: Assertion sym != NULL failed

which i ran into when testing the binary on mint 20.3.

there doesn't seem to be a workaround for this. switching to musl-clang just adds a dependency
on the `musl` package on arch, and it convolutes the dependency build system a bit further.
otherwise, there's not too much lost from switching, since the launcher doesn't have to be
an uber-performance-critical piece of software anyways.

oh well, what can you do.